Changed font from Lane to Round
Changed font for the text elements from Lane to Round.
Changed behavior of the battery (both Pebble and Phone) text elements to remove the '%' symbol when it is at 100%, due to space.

(i.e. if the battery is at 100%, it will just display '100'. Anything less will still have the '%' at the end.)